+++ next/src/components/release/form/schema.ts
@@ -1,23 +1,24 @@
 import z from 'zod'
 
 import i18next from '@/lib/i18n'
-import { TargetModes } from '@/models/questionnaires'
 
 export const schema = z.object({
-  description: z
+  releaseDescription: z
     .string()
     .min(1, { message: i18next.t('release.form.description.required') })
     .max(249, { message: i18next.t('release.form.description.maxLength') }),
-  collectMode: z.enum(TargetModes, {
+  mode: z.enum(['CAWI', 'CAPI', 'CATI'], {
     error: i18next.t('release.form.collectMode.required'),
   }),
-  contexte: z.enum(['menages', 'entreprise'], {
+  context: z.enum(['HOUSEHOLD', 'BUSINESS'], {
     error: i18next.t('release.form.contexte.required'),
   }),
-  pageTempsReponse: z.boolean().default(true),
-  numerotationQuestions: z
-    .enum(['none', 'sequence', 'questionnaire'])
-    .default('sequence'),
+  overrideGenerationParameters: z.object({
+    responseTimeQuestion: z.boolean().default(true),
+    questionNumberingMode: z
+      .enum(['NONE', 'SEQUENCE', 'ALL'])
+      .default('SEQUENCE'),
+  }),
 })
 
 export type FormValues = z.input<typeof schema>